Add option for no transformations (None) in cli
Summary:
Add option for no transformations (None) in cli. Currently there is no way to control transformations from the cli, hence it is always "default".
Ideally we would like transformations argument to be able to get three types of values: None (for no transformations), "default" for default transformations or a dictionary of transformations.
However, as a first step I added the option for "default" (which is also the default) and None (for no transformations).

Reviewed By: talgalili

Differential Revision: D41720439

fbshipit-source-id: 6cc06aca31ee1c5e392af7aed834eb2eeb3b01d1