Fix SettingWithCopyWarning in from_frame (and other places)
Summary:
When the input id column is an int it is changed to be a string. The current operation, in some enviorments, raises
"Casting id column to string"
And then:
```
SettingWithCopyWarning:

A value is trying to be set on a copy of a slice from a DataFrame.
Try using .loc[row_indexer,col_indexer] = value instead

See the caveats in the documentation: https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#returning-a-view-versus-a-copy
```

The current diff is supposed to fix this warning, and also a similar issue in other places.

Reviewed By: SarigT

Differential Revision: D43029054

fbshipit-source-id: ae11abd8f0bdbaf07a5efe09a22e183a011654b0