Sample.from_frame added use_deepcopy argument
Summary:
sample.from_frame has a new argument `use_deepcopy` to decide if changes made to the df inside the sample object would also change the original df that was provided to the sample object. The default is `True` since it's more likely that we'd like to keep the changes inside the sample object to the df contained in it, and not have them spill into the original df. Also added tests to verify this works.

Also added updates to CHANGELOG about this diff and the previous diff, and bumped the current version of balance to 0.3.2.

Reviewed By: SarigT

Differential Revision: D43031296

fbshipit-source-id: 7c0ba9fe02dca8ee6133f04d66aa8df395967744