Add return_df_with_original_dtypes flag to the cli
Summary: This diff adds the ability to have the cli return the sample.df after attempting to restore the dtypes to their original settings. E.g.: an int32 column would be changed into float32 in sample._df, but will then be changed back to in32 before returning from the cli.

Reviewed By: SarigT

Differential Revision: D43043252

fbshipit-source-id: eecefd16b616b03edcfcf823ddd5fb539900b16f