adding sim_data_cbps.csv and making it accessible through load_data
Summary:
This simulated file will be used in a future diff to demonstrate how balance can be used to reproduce the CBPS::CBPS function in R.

Note that I've considered using `importlib.resources`, as described here:
https://stackoverflow.com/a/58936204
But based on the documentation here:
https://docs.python.org/3/library/importlib.resources.html
The API has changed in Python 3.11.
I ended up using `__file__` as described here:
https://dev.to/bowmanjd/easily-load-non-python-data-files-from-a-python-package-2e8g

In the future, we could use importlib.resources and condition the API usage based on the python version. But I suspect this could lead to various lint issues in different python versions.

Reviewed By: SarigT

Differential Revision: D43164088

fbshipit-source-id: e1cbb348d291e87fb91065c96bb9e8d24e3bc2d6