"weight" vs "weights" - more consistant support across functions (preferring weight over weights)
Summary:
Various functions have used "weight" and "weights" inconsistantly across the modules.

For example various implementations for seaborn and plotly functions. Similarly it's an issue with adjustment methods that return a "weights" column even though the default `Sample.from_frame` uses a "weight" column.

This diff aims to make it more consistant by making all the functions just move to use "weight". But also allow "from_frame" to guess that a "weights" column is used for weights.

The decision to go "weight" vs "weights", is that we often use singular for other feature names. E.g.: use "age" and not "ages", "gender" and not "genders", etc.

I came across this issue while trying to debug issues with the plotting functions and had bugs due to this. So decided to just make a consistant decision across the package to minimimze future confusions.

# INTERNAL

This also required some adjustments in 'rake' from the graviton code.

Reviewed By: SarigT

Differential Revision: D44815896

fbshipit-source-id: 997baf9258d3c3d83e50c43551041fad7e55ce71