Add the 'rake' adjustment method
Summary:
Raking is a very well known basic method of adjustment. For details see here:
https://en.wikipedia.org/wiki/Raking

# INTERNAL
This is me finally getting around to move the 'rake' method from graviton to balance.
It still needs more work, but it would be easier to do it now that it's in balance, and out of graviton.

Reviewed By: SarigT

Differential Revision: D45189037

fbshipit-source-id: db62b47bad535d91c7f73c49479015bd1be42cc5