Address deprecated pandas syntax (#46)
Summary:
This PR changes a couple minor deprecated pandas syntax choices.

Passing a set is deprecated. This PR converts the set to a list before indexing with it. This key/value pair in the response doesn't seem to be used elsewhere, but I left it as a set for Chesterton's fence reasons.

Since this warning is surfaced with the first code snippet on the README, it's nice to avoid this warning

Pull Request resolved: https://github.com/facebookresearch/balance/pull/46

Reviewed By: SarigT

Differential Revision: D45722098

Pulled By: talgalili

fbshipit-source-id: a6f83d4b4553dbebbf32dae25d3d633059e086d7