Fix tutorials to include the outcome in the target
Summary:
Due to the recent fix of deprecated pandas code, the outcome of the target population is now properly populated in the functions, but was previously ignored in the notebooks.
This diff fixes it.

Reference to the recent fix:
https://github.com/facebookresearch/balance/commit/7a117a3009db62319e1db7f87363ad7c43368963

Misc: also increased the width of some plots in the quickstart tutorial (for when we later would want to use these images in the docs).

Reviewed By: SarigT

Differential Revision: D45771295

fbshipit-source-id: 9b3d32f511c6bdbdb84f4457ff19df0a4984fc2b