replace np.float with float [5]
Summary:
`numpy.float` has been deprecated and removed in numpy-1.20.0. It was an alias to the builtin `float` and should be safe to replace with `float`.

The whole change was mechanical with the following one-liner:
```
fbgr -sl '\bnp\.float\b' | xargs perl -pi -e 's,\bnp\.float\b,float,g'
```

Differential Revision: D46609273

fbshipit-source-id: 5b9783bce99560630a825b1035672058981be873